chore(ci): split upgrade from install
